+++ src/actions/functions/toAction.ts
@@ -1,7 +1,7 @@
-import type { Fn, PrimitiveObject } from '@bemedev/types/lib/types/types';
+import type { PrimitiveObject } from '@bemedev/types/lib/types/types';
 import type { ActionConfig, ActionMap, ActionResult } from '~actions';
-import type { EventsMap, PromiseeMap, ToEvents } from '~events';
-import { isDescriber } from '~types';
+import type { EventsMap, PromiseeMap } from '~events';
+import { isDescriber, type FnR } from '~types';
 import { reduceFnMap } from '~utils';
 
 export type ToAction_F = <
@@ -14,7 +14,7 @@ export type ToAction_F = <
   promisees: P,
   action: ActionConfig,
   actions?: ActionMap<E, P, Pc, Tc>,
-) => Fn<[Pc, Tc, ToEvents<E, P>], ActionResult<Pc, Tc>> | undefined;
+) => FnR<E, P, Pc, Tc, ActionResult<Pc, Tc>> | undefined;
 
 /**
  * Converts an ActionConfig to a function that can be executed with the provided eventsMap and promisees.
@@ -22,9 +22,7 @@ export type ToAction_F = <
  * @param promisees of type {@linkcode PromiseeMap}, the promisees map to use for resolving the action.
  * @param action of type {@linkcode ActionConfig}, action configuration to convert.
  * @param actions of type {@linkcode ActionMap}, The actions map containing functions to execute.
- * @returns a {@linkcode Fn} function that executes the action or undefined if not found.
  *
- * @see {@linkcode ToEvents}
  * @see {@linkcode PrimitiveObject}
  * @see {@linkcode ActionResult}
  * @see {@linkcode reduceFnMap}